Introduce ShoppingCart entity and implement corresponding model, service, controller. Refine the implementation of other entities.

+++ src/main/java/com/example/demo/request/AddProductRequest.java
@@ -1,13 +1,15 @@
 package com.example.demo.request;
 
-import lombok.*;
+import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 public class AddProductRequest {
 
     private String name;
 
-    private Double price;
+    private BigDecimal price;
 
     private String description;
 }
